# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminUserOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "active"
        ],
        "title": "AdminUserOut",
        "type": "object"
      },
...
      "ColumnView": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "kind",
          "date"
        ],
        "title": "ColumnView",
        "type": "object"
      },
...
      "EventOut": {
        "properties": {
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "changed_labels": {
            "items": {
              "type": "string"
            },
            "title": "Changed Labels",
            "type": "array"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "dossier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dossier Id"
          },
          "dossier_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dossier Number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "insured_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insured Name"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "membership_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Membership Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "notification_class": {
            "$ref": "#/components/schemas/NotificationClass"
          },
          "occurrence_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurrence Id"
          },
          "workflow_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Key"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "required": [
          "id",
          "kind",
          "notification_class",
          "detected_at",
          "workflow_key",
          "workflow_name",
          "dossier_id",
          "dossier_number",
          "insured_name",
          "occurrence_id",
          "membership_id",
          "changed_fields",
          "changed_labels",
          "message"
        ],
        "title": "EventOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LastRunView": {
        "properties": {
          "baseline": {
            "title": "Baseline",
            "type": "boolean"
          },
          "changed": {
            "title": "Changed",
            "type": "integer"
          },
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "details_failed": {
            "title": "Details Failed",
            "type": "integer"
          },
          "left": {
            "title": "Left",
            "type": "integer"
          },
          "returned": {
            "title": "Returned",
            "type": "integer"
          },
          "rows_seen": {
            "title": "Rows Seen",
            "type": "integer"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PollStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "status",
          "baseline",
          "rows_seen",
          "created",
          "returned",
          "changed",
          "left",
          "details_failed"
        ],
        "title": "LastRunView",
        "type": "object"
      },
...
      "NotificationClass": {
        "description": "How a workflow's activity reaches employees.\n\n``ACTION`` events can create unread alerts; ``INFORMATIONAL`` events are\nonly shown in history/summaries; ``SILENT`` workflows are synchronized\nfor search and detail context and emit no events.",
        "enum": [
          "ACTION",
          "INFORMATIONAL",
          "SILENT"
        ],
        "title": "NotificationClass",
        "type": "string"
      },
...
      "PollRunView": {
        "properties": {
          "baseline": {
            "title": "Baseline",
            "type": "boolean"
          },
          "changed": {
            "title": "Changed",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "details_failed": {
            "title": "Details Failed",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "left": {
            "title": "Left",
            "type": "integer"
          },
          "notifications_created": {
            "title": "Notifications Created",
            "type": "integer"
          },
          "pages_seen": {
            "title": "Pages Seen",
            "type": "integer"
          },
          "returned": {
            "title": "Returned",
            "type": "integer"
          },
          "rows_seen": {
            "title": "Rows Seen",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PollStatus"
          },
          "sync_run_id": {
            "title": "Sync Run Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sync_run_id",
          "status",
          "started_at",
          "completed_at",
          "baseline",
          "rows_seen",
          "pages_seen",
          "created",
          "returned",
          "changed",
          "left",
          "notifications_created",
          "details_failed",
          "error"
        ],
        "title": "PollRunView",
        "type": "object"
      },
      "PollStatus": {
        "enum": [
          "COMPLETE",
          "PARTIAL",
          "AUTH_REQUIRED",
          "FAILED"
        ],
        "title": "PollStatus",
        "type": "string"
      },
      "Role": {
        "enum": [
          "ADMIN",
          "EMPLOYEE"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "UserActiveIn": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          }
        },
        "required": [
          "active"
        ],
        "title": "UserActiveIn",
        "type": "object"
      },
      "UserCreateIn": {
        "properties": {
          "display_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "password": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "default": "EMPLOYEE"
          },
          "username": {
            "maxLength": 150,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "display_name",
          "password"
        ],
        "title": "UserCreateIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowConfigIn": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "notification_class": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationClass"
              },
              {
                "type": "null"
              }
            ]
          },
          "rules_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowRulesStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "WorkflowConfigIn",
        "type": "object"
      },
      "WorkflowDetailView": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnView"
            },
            "title": "Columns",
            "type": "array"
          },
          "filter_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filter Label"
          },
          "primary_date_labels": {
            "items": {
              "type": "string"
            },
            "title": "Primary Date Labels",
            "type": "array"
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "title": "Recent Events",
            "type": "array"
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/PollRunView"
            },
            "title": "Recent Runs",
            "type": "array"
          },
          "workflow": {
            "$ref": "#/components/schemas/WorkflowView"
          }
        },
        "required": [
          "workflow",
          "filter_label",
          "primary_date_labels",
          "columns",
          "recent_runs",
          "recent_events"
        ],
        "title": "WorkflowDetailView",
        "type": "object"
      },
      "WorkflowRulesStatus": {
        "description": "Whether employees have approved a workflow's business policy.\n\n``UNCONFIRMED`` workflows are catalogued but inactive. ``CAPTURE_DERIVED``\nworkflows run with the conservative queue-membership policy derived from\nthe ScriptScrap captures and are shown as \"À valider sur site\" until the\nagency validates them. ``CONFIRMED`` workflows were validated by the\nagency. Promoting a rule never changes the event engine.",
        "enum": [
          "UNCONFIRMED",
          "CAPTURE_DERIVED",
          "CONFIRMED"
        ],
        "title": "WorkflowRulesStatus",
        "type": "string"
      },
      "WorkflowView": {
        "properties": {
          "active_count": {
            "title": "Active Count",
            "type": "integer"
          },
          "baseline_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Completed At"
          },
          "by_work_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Work Status",
            "type": "object"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_poll_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Poll At"
          },
          "last_poll_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PollStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LastRunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "needs_site_validation": {
            "title": "Needs Site Validation",
            "type": "boolean"
          },
          "notification_class": {
            "$ref": "#/components/schemas/NotificationClass"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "rules_status": {
            "$ref": "#/components/schemas/WorkflowRulesStatus"
          },
          "unread_changed": {
            "title": "Unread Changed",
            "type": "integer"
          },
          "unread_new": {
            "title": "Unread New",
            "type": "integer"
          },
          "view_id": {
            "title": "View Id",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "category",
          "route",
          "view_id",
          "notification_class",
          "rules_status",
          "needs_site_validation",
          "enabled",
          "baseline_completed_at",
          "last_poll_at",
          "last_success_at",
          "last_poll_status",
          "last_error",
          "active_count",
          "unread_new",
          "unread_changed",
          "by_work_status"
        ],
        "title": "WorkflowView",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Portail RMA",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "list_users_api_v1_admin_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminUserOut"
                  },
                  "title": "Response List Users Api V1 Admin Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Users",
        "tags": [
          "operations",
          "admin"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create User",
        "tags": [
          "operations",
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "patch": {
        "operationId": "set_user_active_api_v1_admin_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActiveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set User Active",
        "tags": [
          "operations",
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/workflows/{key}": {
      "patch": {
        "description": "Enable/disable a queue, promote its rules (À valider sur site -> confirmé) or change\nhow its activity reaches employees. The event engine itself never changes.",
        "operationId": "update_workflow_api_v1_admin_workflows__key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workflow",
        "tags": [
          "operations",
          "admin"
        ]
      }
    },
...
    "/api/v1/workflows/{key}": {
      "get": {
        "operationId": "workflow_detail_api_v1_workflows__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workflow Detail",
        "tags": [
          "workspace"
        ]
      }
    },
...
  }
}